// src/types/types.ts
type AuthorCountByYear = {
    year: number;
    works_count?: number;
    cited_by_count: number;
}

type ExternalIds = {
    openalex?: string;
    orcid?: string;
    scopus?: string;
    twitter?: string;
    wikidata?: string;
    issn_l?: string;
    issn?: string[];
    mag?: number;
    fatcat?: string;
}

type SummaryStats = {
    "2yr_mean_citedness": number;
    "h_index": number;
    "i10_index": number;
}
// ...
enum PossibleInstitutionRole {
    FUNDER = 'funder',
    INSTITUTION = 'institution',
    PUBLISHER = 'publisher',
}

type InstitutionRole = {
    role: PossibleInstitutionRole;
    id: string;
    work_count: number;
}
// ...
type FunderResponse = {
    alternate_titles: string[];
    cited_by_count: number;
    country_codes: Iso3166Alpha2CountryCode[];
    counts_by_year: AuthorCountByYear[];
    created_date: string[];
    description: string;
    display_name: string;
    grants_count: number;
    homepage_url: string;
    id: string;
    ids: ExternalIds;
    image_thumbnail_url: string;
    image_url: string;
    roles: InstitutionRole[];
    summary_stats: SummaryStats;
    updated_date: string;
    works_count: number;

}